Call lifecycle callbacks on registered elements
The hack to unregister elements requires that the lifecycle methods
associated with the elements don’t change, so I delegate to versions of
the methods prefixed with _ from the default prototype’s hooks. This
does pick up changes to the prototype correctly. I’m still a bit worried
about what happens after unregistering…