Fix errors in README
If an arrow function is turned into a block it does not implicitly return as is
the case when the arrow function is not a block. Using the old example code in
the README would result in a DOM level error originating at
partial-observation-widget.js#L158 caused by an attempt to insert a `null`.