Add etch.stateless for creating stateless components
Often, a componnet only needs to take in props/children and render
something, and does not maintain any internal state. In those instances,
it can be useful to define the component as a function of props and
children.

This adds `etch.stateless(renderFn)` that implements a component that
renders `renderFn` with props and children during `initialize` and
`update`.

Signed-off-by: Katrina Uychaco <kuychaco@github.com>